refactor: Pawn 객체 하나의 파일로 관리하기
