+++ src/main/java/chess/domain/pieces/Piece.java
@@ -37,21 +37,4 @@ public abstract class Piece {
     public boolean isBlackTeam() {
         return team.isBlackTeam();
     }
-
-    @Override
-    public boolean equals(final Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        Piece piece = (Piece) o;
-        return team == piece.team;
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(team);
-    }
 }
